+++ server/models/CakeFlavor.js
@@ -2,7 +2,7 @@ const { DataTypes } = require('sequelize');
 const { sequelize } = require('../config/database');
 
 const CakeFlavor = sequelize.define('CakeFlavor', {
-    tenantId: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 1 },
+    tenantId: { type: DataTypes.BIGINT, allowNull: false, defaultValue: 1 },
     name: { type: DataTypes.STRING, allowNull: false },
     isActive: { type: DataTypes.BOOLEAN, defaultValue: true },
 }, {
