+++ server/models/Ingredient.js
@@ -2,7 +2,7 @@ const { DataTypes } = require('sequelize');
 const { sequelize } = require('../config/database');
 
 const Ingredient = sequelize.define('Ingredient', {
-    tenantId: { type: DataTypes.INTEGER, allowNull: true },
+    tenantId: { type: DataTypes.BIGINT, allowNull: true },
     name: { type: DataTypes.STRING(120), allowNull: false },
     unit: { type: DataTypes.STRING(20), allowNull: false }, // kg, g, pza, lt
     stock: { type: DataTypes.DECIMAL(10, 2), defaultValue: 0.00 },
